+++ CHANGES.md
@@ -18,6 +18,7 @@ Improvements 🙌:
  - Room list improvements (paging)
  - Fix quick click action (#3127)
  - Get Event after a Push for a faster notification display in some conditions
+ - Always try to retry Http requests in case of 429 (#1300)
 
 Bugfix 🐛:
  - Fix bad theme change for the MainActivity